I made some changes while trying to read in a QIF file from my version of quicken (Quicken Premier 2018, R11.20). The parsing issues I had:
* The first line in my file had a trailing space, "!Type:Bank ".  Adding .strip() to remove trailing whitespace.

* My .qif file had commas in the thousands place:  "T-2,347.17".  Added ".replace(',','')" to remove commas.
** Note: My file only had Bank transactions, I added them everywhere it parsed in a Decimal item, though, to be through.

* My .qif file had 'U' codes as well as 'T' codes.
** Per https://en.wikipedia.org/wiki/Quicken_Interchange_Format the U code seems equivalent to T codes.

* My .qif file had dates like D11/15'13 which should be converted to 2013-11-15
** I'm concerned this might break handling of older files, which might be Ddd/mm/'yy, instead of Dmm/dd/'yy ??

* I had transactions with a value of 0.0.  This was failing a check for "not val".  I changed it to just check for None (the default value).

The tests are "mostly" working, but it IS failing when writing out/comparing some of the "older" dates, and also when writing back out the U codes.
I'm not very familiar with QIF and the best way to fix these.
Perhaps add another field (of type U) to the Transaction class?
These should be fixed before merging to master, but I wanted to document these fixes before losing them.